requirements.txt
Flask
openai
mysql-connector-python
python-dotenv
Flask-Limiter